refactor model store
