+++ onnxscript/rewriter/ort_fusions/fuse_xformers_test.py
@@ -6,7 +6,7 @@ import unittest
 
 import onnxscript.optimizer
 from onnxscript.rewriter.ort_fusions._core import fuse_xformers
-from onnxscript.rewriter.ort_fusions._smollm_1 import smollm_test_1
+from onnxscript.onnxscript.rewriter.ort_fusions.models._smollm_1 import smollm_test_1
 from onnxscript.rewriter.ort_fusions._test_utils import assert_allclose, ort_run
