+++ onnxscript/rewriter/ort_fusions/models/_rotary_embedding_models.py
