+++ onnxscript/rewriter/ort_fusions/models/_smollm_2.py
