+++ onnxscript/rewriter/ort_fusions/rms_normalization_test.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import unittest
 
 import onnxscript.optimizer
-from onnxscript.rewriter.ort_fusions._smollm_1 import smollm_test_1
+from onnxscript.onnxscript.rewriter.ort_fusions.models._smollm_1 import smollm_test_1
 from onnxscript.rewriter.ort_fusions._test_utils import assert_allclose, ort_run
 from onnxscript.rewriter.ort_fusions.rms_normalization import fuse_rms_normalization
